+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.7.0] - 2021-01-19
+
 ### Added
 
 - Adding label value validation for `Cluster` CR for non-version labels.
@@ -173,7 +175,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release.
 
-[Unreleased]: https://github.com/giantswarm/aws-admission-controller/compare/v2.6.0...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-admission-controller/compare/v2.7.0...HEAD
+[2.7.0]: https://github.com/giantswarm/aws-admission-controller/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/giantswarm/aws-admission-controller/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/giantswarm/aws-admission-controller/compare/v2.4.1...v2.5.0
 [2.4.1]: https://github.com/giantswarm/aws-admission-controller/compare/v2.4.0...v2.4.1